add pagination tests (incomplete) and seperate the tests from navbar tests
